Provide default value for DRONE_COMMIT_SHA
Prior to this commit, when running `drone exec` locally to trigger the
drone pipeline, there was no commit value present in the
DRONE_COMMIT_SHA environment variable. This would cause the pipeline to
fail.

By providing a default of NO_SHA, the pipeline will succeed and any
build artifacts will be not be given names, tags or labels that will
make them seem like they are production-ready.