Increase number of retries and backoff period
Priot to this change, the connection between the test suite and the
monserve container would often not get established when starting the
tests.

This change doubles both the number of retries to do and the period
in between each retry to give less-powerful machine a better chance to
have the functional tests establish their connection.