Add test to check version subommand output
Prior to this command, if the ldflags would have changed, there wasn't
anything in place to ensure that the version variable inside an app
would still get changed.

This change provides a simple test to check that the version subcommand
outputs the version that has been calculated within the make command.