Use package instead of main.go in goreleaser
By providing explicitly the main.go file to the goreleaser, only the
files that are referenced by main.go become compiled. This meant that
the version.go file within the monserve pkg was not being compiled, so
there was no version subcommand being added to the root command.

This change provides the package using cmd/monserve to the goreleaser
for both monserve and moncli, to avoid this error in future.