Remove install and clean rules from makefile
Prior to this change, when running `make` without any arguments, the
binaries would be installed on the local machine and then the build dir
would be deleted. With the move to CI/CD, there is no need to run the
install rule and the clean rule just deletes the compiled binaries,
which may be required for other parts of the CI process.