Remove ldflags and add dirty to version
There is no version variable within the tests, so not need to set the
version using ldflags.

Prior to this change, there would be no difference in image tag between
functional tests on an unchnaged tagged repo, vs a dirty repo (where
there is WIP in the repo after a tag).

This change adds the `--dirty` flag to the git command, to report
whether the version has uncommited work.