Refactorpostgrestestmakefile (#32)
* Remove ldflags and add dirty to version

There is no version variable within the tests, so not need to set the
version using ldflags.

Prior to this change, there would be no difference in image tag between
functional tests on an unchanged tagged repo, vs a dirty repo (where
there is WIP in the repo after a tag).

This change adds the `--dirty` flag to the git command, to report
whether the version has uncommitted work.

* Use more variables within Makefile for more configurability

* Always `docker-compose down` before functional tests

Prior to this change, if there was some kind of failure when running
the tests, trying to run the tests again could have strange effects from
the previous try.

By running `docker-compose down` first, any previous network will be
destroyed, leaving a state ready to start some new tests.

* Rename make rule build-tests -> build-test-binary

prior to this change it wasn't explicitly clear what the build-tests
would be doing. It could have been building an image or a binary.

This change makes this explicit.