Change storage interface to not take pointer args
Prior to this change, the UpdateAccount method of the Sotrage interface
had pointers for arguments. This was unnecessary as the value was never
changed and it is only a small data structure.

This change alters the method signature and updates all relevent uses
of the interface to work in the same way.