Move UpdateAccount business logic into models package
Prior to this change, the business logic of the updating accounts, i.e.
validation etc., was witihn the storage package.

This change creates a models package which will contain the business
logic, although I am hoping that the package will be renamed at a later
stage.

The tests for storage have had to change, so the models package has some
tests, which in turn have meant that the storagetest.Storage has had to
become a little more advanced, now being a mock rather than just a stub.
This means that when there is a storage method executed with an ID for
an account, this ID is recorded and can then be checked against in the
tests.