Show date of latest balance in accounts balances cmd
When the model of the data was changed to be delta balances instead of
absolute balances, the logic of moncli was altered, including showing
the date of the query in the "latest balance" column, rather than the
date of the latest balance that contributes to the account total.

This change alters the code to show the date of the latest balance that
contributed to the total.

The signature of accountBalanceAtTime(storage.Storage,
storage.Account,time.Time) (balance.Balance, error)
has been changed to accountBalancesAtTime(store storage.Storage,
a storage.Account, at time.Time) (storage.Balances, error)
so that it is more reusable, with less responsibility.