Remove repetition in balance sorting functions
Prior to this change, there was a similar boilerplate in each of the
sorting functions which would call sort.Slice with some parameters.
Each time, the boilerplate would cause the author to require using the
slice twice within the function. Although this is quite simple, it could
cause issues further down the line if the sorting were to become more
complex. This boilerplate also made adding new sorting orders
cumbersome.

This change adds a type, accountBalanceComparison, which is a function
that takes 2 AccountBalance objects and returns true if they're in
order. This type can now be used as an argument to the reusable
function, which used to be boilerplate, to generate all of what we need
to do the sorting.