balancesoftdelete: Add deleted column to balances
When calling DeleteBalance, `time.Now()` is used to determine the
current time in the background. This is the same behaviour as with when
deleting an Account, so has been determined to be acceptible for the
moment. They can both be alterred in the future when a decision has been
made about how best to do this.