balancesoftdelete: Extract query fields into vars
Priot to this commit, when creating the tables in the postgres storage,
the columns were hardcoded into the query string, this change extracts
them into an fmt.Sprintf function that uses the constants.

This way, if we want to change the column names, there will only be one
place that we need to chane them and the difference will propagate.