goreleaser-docker: generate dockerfiles from template
Prior to this commit, there was a single Dockerfile that was used for
each binary. This would mean that for each binary, it would be called
'/app' within the container. Although this isn't much of an issue, it
means that the process name was wrong.

When trying to use the single dockerfile for every binary within
goreleaser, this was causing issues.

There is now a Dockerfile.template which can be used to generate a
simple Dockerfile.APP_NAME for any app within the repo.