client-requestfactory: Use http.Request in methods
Prior to this change, in some of the methods of the client, the helper
functions from the http package that would use the DefaultClient were
used to make requests. Helper functions such as `http.Get`.

This change makes each use of the helper functions create a Request and
then use that with a client to request data from the server.

Use of the DefaultClient can cause issues with timeouts as it would
hang forever, rather than terminating if not successful for a certain
period of time. This will take up resources and would, perhaps, not end
up surfacing any issues if left open and unnoticed for long periods of
time.