client-requestfactory: Change Client to struct
Prior to this commit, the Client was an alias for a string, which was
the url of the server.

This change turns the Client to a struct with, currently, just a single
field of URL.

This allows the cient to be extended to have other fields a lot more
easily.