Client Request Factory (#59)
* client-requestfactory: Use http.Request in methods
Prior to this change, in some of the methods of the client, the helper
functions from the http package that would use the DefaultClient were
used to make requests. Helper functions such as `http.Get`.

This change makes each use of the helper functions create a Request and
then use that with a client to request data from the server.

Use of the DefaultClient can cause issues with timeouts as it would
hang forever, rather than terminating if not successful for a certain
period of time. This will take up resources and would, perhaps, not end
up surfacing any issues if left open and unnoticed for long periods of
time.

* client-requestfactory: Change Client to struct
Prior to this commit, the Client was an alias for a string, which was
the url of the server.

This change turns the Client to a struct with, currently, just a single
field of URL.

This allows the cient to be extended to have other fields a lot more
easily.

* newRequest method on client
This creates a new request that is configured to call the server at the
client's configured url